% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate.R
\name{between2}
\alias{between2}
\title{Does numeric vector fall in a range}
\usage{
between2(x, left = -Inf, right = Inf, leq = TRUE, geq = TRUE,
  na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector of values}

\item{left, right}{Bountary values}

\item{geq, leq}{Use \code{>=} (\code{<=}) else use \code{>} (\code{<}) for testing
the boundaries.}

\item{na.rm}{If \code{TRUE} return \code{FALSE} for \code{NA} or \code{NaN}}
}
\value{
A logical vector that is \code{TRUE} if the value falls in the interval.
}
\description{
This is a shortcut for \code{x > lower & x < upper} and variants with equalities.
This function is more flexible than \code{\link[dplyr]{between}}.
}

