% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_cases.R
\name{complete_cases}
\alias{complete_cases}
\alias{complete_cases_}
\title{Keep complete cases of a table}
\usage{
complete_cases(.data, ...)

complete_cases_(.data, ..., .dots)
}
\arguments{
\item{.data}{A \code{tbl}. Currently only in-memory data frames are supported.}

\item{...}{Comma separated list of unquoted expressions. This uses the same
semantics as \code{\link[dplyr]{select}}. You can select a subset of
variables to check for missing values.}

\item{.dots}{Used to work around non-standard evaluation.
See \code{vignette("nse")} for details.}
}
\value{
The table with rows containing \code{NA} values dropped.
}
\description{
Filter a table by keeping only rows with no missing values.
This is more tidyverse friendly version of \code{complete.cases}, and
saves writing potentially long \code{\link[dplyr]{filter}} expressions.
}
\references{
\url{http://stackoverflow.com/a/37031161/227406}
}
\author{
Jan Schultz
}
