% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equicorr.R
\name{equicorr}
\alias{equicorr}
\title{Create a equicorrelated matrix}
\usage{
equicorr(k, rho = 0, scale = NULL)
}
\arguments{
\item{k}{An integer with the matrix dimension.}

\item{rho}{The correlation coefficient for off-diagonal entries.}

\item{scale}{If \code{NULL} a correlation matrix is returned.
If a vector of standard deviations, then a covariance matrix is returned.}
}
\value{
A correlation or covariance matrix (if \code{scale} is not-null).
}
\description{
Create an correlation matrix in which all off-diagonal
elements are equal,
\deqn{
\begin{bmatrix}
1 & \rho & \dots & rho \\
\rho & 1 & \dots & rho \\
\vdots & \vdots & \ddots & \vdots \\
\rho & \rho & \dots & 1
\end{bmatrix}
}
}
