% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{fct_map}
\alias{fct_map}
\title{Transform levels of a factor with a function}
\usage{
fct_map(f, .f, ...)
}
\arguments{
\item{f}{A factor}

\item{.f}{A function that takes the current levels as its first argument.}

\item{...}{Arguments passed to \code{.f}.}
}
\value{
A factor vector with the values of \code{f} and transformed levels.
}
\description{
Change the factor levels by a function of their current values.
}

