% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{fct_remove}
\alias{fct_remove}
\title{Remove levels from a factor}
\usage{
fct_remove(f, lvls)
}
\arguments{
\item{f}{A factor}

\item{lvls}{A character vector of levels to remove from the factor, and replace with \code{NA} values.}
}
\value{
A factor vector
}
\description{
Remove levels from a factor
}
\examples{
x <- factor(c("apple", "bear", "banana", "deer"))
fct_remove(x, c("apple", "banana"))
}
\seealso{
\code{\link[forcats]{fct_remove}} can also remove levels.
}

