% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{fct_seq}
\alias{fct_seq}
\title{Transform levels of a factor with a function of their index}
\usage{
fct_seq(f, .f = "\%d", ...)
}
\arguments{
\item{f}{A factor}

\item{.f}{If a character vector, a \code{\link{sprintf}} pattern in which the
only argument will be the order of the factor levels. If a function, then
a function in which the first argument is the order of the factor levels.}

\item{...}{Arguments passed to \code{.f} if it is a function.}
}
\value{
A factor vector with the values of \code{f} and transformed levels.
}
\description{
Change the factor levels by a function or pattern based on their current
order, .e.g. 1, 2, 3, ....
}

