% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_na.R
\name{filter_na}
\alias{filter_na}
\alias{filter_na_}
\title{Filter rows with missing values}
\usage{
filter_na(.data, ..., .all = FALSE, .finite = FALSE)

filter_na_(.data, ..., .dots, .all = FALSE, .finite = FALSE)
}
\arguments{
\item{.data}{A \code{tbl}}

\item{...}{Comma separated list of unquoted expressions. Select variables
which to test for missing values using the same semantics as
\code{\link[dplyr]{select}}.}

\item{.all}{If \code{TRUE}, then remove rows in which all columns are
missing, else select rows in which any column has a missing value.}

\item{.finite}{For numeric columns, filter using \code{\link{is.finite}},
which will drop rows with \code{NaN}, \code{-Inf}, and \code{Inf} values in
addition to \code{NA}.}

\item{.dots}{use \code{filter_na_()} to do standard evaluation. See
\code{vignette("nse")} for details.}
}
\value{
An object of the same class as \code{.data}.
}
\description{
Remove rows with missing values in variables that are selected.
}
