% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{glob}
\alias{glob}
\title{Select helper to select variables based on name glob pattern.}
\usage{
glob(pattern, ignore.case = TRUE, vars = current_vars())
}
\arguments{
\item{pattern}{Globbing pattern}

\item{ignore.case}{If \code{TRUE}, the default, ignores case when matching names.}

\item{vars}{A character vector of variable names.
When called from inside \code{select()} these are automatically set
to the names of the table.}
}
\value{
An integer vector given the position of the matched variables.
}
\description{
Like \code{\link[dplyr]{matches}}, this function selects variables by by name.
However, it uses wildcard, aka globbing patterns instead of regular expressions.
}
\details{
Use \code{?} to match a single character, and \code{*} to match any number of
characters, including none. The pattern is anchored at the start and end of the
string, meaning it must start at the start, and end at the the end.

This function uses \code{\link[utils]{glob2rx}} to convert the globbing pattern
to a regex,
}

