% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubbish.R
\name{\%==\%}
\alias{\%==\%}
\title{Test for equality with NA, NaN}
\usage{
x \%==\% y
}
\arguments{
\item{x, y}{Values to test}
}
\value{
A logical vector with the test results.
}
\description{
Like \code{x == y}, but not returning \code{NA} in tests involving
\code{NaN} or \code{NA}.
}
\examples{
x <- c(1, NA)
y <- c(1, NA)

}

