% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{has_names}
\alias{has_names}
\title{Check that an object has names}
\usage{
has_names(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
Check that an object has names
}
\examples{
has_names(c(1, 2, 3))
has_names(c(a = 1, b = 2, c = 3))
}

