% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{has_no_missing_names}
\alias{has_no_missing_names}
\title{Check that all names in an object are not-missing}
\usage{
has_no_missing_names(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
Check that all names in an object are not-missing
}
\examples{
has_no_missing_names <- list(1, b = 1)
has_no_missing_names <- list(a = 1, b = 1)
}

