% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{has_unique_names}
\alias{has_unique_names}
\title{Check that an object has unique names}
\usage{
has_unique_names(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
Check that an object has unique names
}
\examples{
has_unique_names(list(a = 1, b = 2))
has_unique_names(list(a = 1, a = 2))
}

