% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{is_formula2}
\alias{is_formula2}
\alias{is_formula1}
\title{Is a one- or two-sided formula?}
\usage{
is_formula2(x)

is_formula1(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Functions to test whether an object is a one- or two-sided formula.
}
\examples{
is_formula2(y ~ x)
is_formula2(~ x)
is_formula1(y ~ x)
is_formula1(~ x)
}
\seealso{
\code{\link[purrr]{is_formula}} in \pkg{purrr} and \code{\link[lazyeval]{is_formula}} in \pkg{lazyeval} both test objects for formula.
}
