% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_keys.R
\name{is_key}
\alias{is_key}
\alias{is_key_}
\title{Check if variables are distinct}
\usage{
is_key(.data, ...)

is_key_(.data, ..., .dots)
}
\arguments{
\item{.data}{a tbl}

\item{...}{Variables to check whether they are keys}

\item{.dots}{Used to work around non-standard evaluation.
See \code{vignette("nse")} for details.}
}
\value{
\code{TRUE} if the variables jointly identify all
  rows, and \code{FALSE} otherwise.
}
\description{
Check if a combination of variable(s) uniquely identify
rows of a table.
}
