% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_names.R
\name{map_names}
\alias{map2_names}
\alias{map_names}
\title{Set names in a vector with a function}
\usage{
map_names(.x, .f, ...)

map2_names(x, .f, ...)
}
\arguments{
\item{.f}{Function to use to name the variables.
This is converted to a function using \code{\link[purrr]{as_function}}.
In \code{set_names_map}, the first argument of \code{.f} is
\code{names(x)}. In \code{set_names_map2}, the first argument of \code{.f}
is the \code{names(x)} and the second argument is \code{x}.
Pass additional arguments to \code{.f} with \code{...}.
It should return a character vector of the same length as the
names of \code{x}.}

\item{...}{Arguments passed to the function \code{.f}.}

\item{x}{Vector to name}
}
\value{
The vector \code{x} with new names
}
\description{
Set names in a vector by applying a function to the original names.
}
\examples{
require("stringr")
map_names(c("a" = 1, "b" = 2, "c" = 3), str_to_upper)
map2_names(c("a" = 1, "b" = 2, "c" = 3),
             function(x, y) str_c(x, y, sep = "-"))
}

