% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummies.R
\name{model_df}
\alias{model_df}
\title{Transform a data frame using a formula}
\usage{
model_df(.data, formula, append = FALSE, na.action = "na.pass")
}
\arguments{
\item{.data}{A data frame}

\item{formula}{A formula}

\item{append}{Add columns to \code{.data}. If \code{FALSE}, then return a new
data frame.}

\item{na.action}{The \code{\link{na.action}} to use when creating the \code{model.matrix}.}
}
\value{
A data frame with the transformations specified in the formula.
}
\description{
This uses \code{model.matrix} formula to transform a data frame.
}

