% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junk.R
\name{na2f}
\alias{na2f}
\title{Replace NA with FALSE}
\usage{
na2f(x)
}
\arguments{
\item{x}{A logical vector}
}
\value{
The vector \code{x} with any \code{NA} values replace with \code{FALSE}
}
\description{
Replace NA with FALSE
}
\examples{
x <- c(1, 2, NA) < 5
x
na2f(x)
}

