% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_names.R
\name{nmap}
\alias{nmap}
\alias{nmap_chr}
\alias{nmap_dbl}
\alias{nmap_df}
\alias{nmap_int}
\alias{nmap_lgl}
\title{Map over values and names simultaneoulsy}
\usage{
nmap(.x, .f, ...)

nmap_lgl(.x, .f, ...)

nmap_dbl(.x, .f, ...)

nmap_int(.x, .f, ...)

nmap_int(.x, .f, ...)

nmap_chr(.x, .f, ...)

nmap_df(.x, .f, ...)
}
\arguments{
\item{.x}{A named vector}

\item{.f}{A function.}

\item{...}{Additional arguments passed to \code{.f}}
}
\value{
\code{nmap} returns a list, \code{nmap_lgl} returns a logical vector,
  \code{nmap_dbl} a numeric vector, \code{nmap_int} an integer vector,
  and \code{nmap_df} a data frame.
}
\description{
These functions are helper functions for common case of
iterating over both values and names of a vector in parallel.
They are equivalent to \code{map2(.x, names(.x), .f, ...)}.
}

