% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{pkg_check_examples}
\alias{pkg_check_examples}
\title{Check a package documentation for examples}
\usage{
pkg_check_examples(dir, quiet = FALSE)
}
\arguments{
\item{dir}{Root directory of a package}

\item{quiet}{If \code{TRUE}, then no message is displayed.}
}
\value{
A character vector of Rd files which
do not have an example code. This function is primarily called
for its side effect of printing the list of files.
}
\description{
Check a package documentation for examples
}
