% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{prop}
\alias{prop}
\alias{prop_}
\title{Proportion of observations by group}
\usage{
prop(x, ..., wt = NULL, sort = FALSE)

prop_(x, vars, wt = NULL, sort = FALSE)
}
\arguments{
\item{x}{A tbl}

\item{wt}{If omitted, will the number of rows. If specified, will perform a "weighted" proportion by summing the (non-missing) values of variable \code{wt}.}

\item{sort}{If \code{TRUE}, sort output in descending order of \code{n}.}

\item{vars, ...}{Variables to group by}
}
\value{
A tbl with the grouping variables and column \code{.prop} with the
  proportion in each group.
}
\description{
\code{prop} calculates the proportion of observations by group.
This is analagous to \code{\link[dplyr]{count}} which counts the number of
rows by group. Alternatively, \code{\link[dplyr]{count}} is to \code{\link[base]{table}}
as \code{prop} is to \code{\link[base]{prop.table}}.
}

