% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_names.R
\name{recode_names}
\alias{recode_names}
\title{Recode names of a vector}
\usage{
recode_names(x, .f, ...)
}
\arguments{
\item{x}{A vector}

\item{.f}{A named character vector or something that
can be coerced into a function. If it is a function
it must return a character vector the same length as
\code{names(x)}.}

\item{...}{Arguments passed to \code{.f} if it is a function.}
}
\value{
The vector \code{x} with new names.
}
\description{
Recode names of a vector
}
