% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_map}
\alias{rename_map}
\title{Rename tables with a function}
\usage{
rename_map(.data, .f, ...)
}
\arguments{
\item{.data}{A data frame}

\item{.f}{A function that takes the names of \code{.data} as its first argument.}

\item{...}{Additional arguments to pass to \code{.f}}
}
\description{
Rename tables with a function
}

