% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_replace}
\alias{rename_replace}
\alias{rename_replace_all}
\title{Rename tables with a regular expression}
\usage{
rename_replace(.data, pattern, replacement)

rename_replace_all(.data, pattern, replacement)
}
\arguments{
\item{.data}{A data frame}

\item{pattern, replacement}{Pattern and replacement regular expressions.
See \code{\link[stringr]{str_replace}}.}
}
\description{
Use a regular expression to rename tables.
The function \code{rename_replace} applies \code{\link[stringr]{str_replace}} to the names of the table,
while \code{rename_replace_all} applies \code{\link[stringr]{str_replace_all}}.
}
