% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_seq}
\alias{rename_seq}
\title{Rename tables with sequentially numbered variables}
\usage{
rename_seq(.data, .f = "Var\%d", ...)
}
\arguments{
\item{.data}{A data frame}

\item{.f}{Either a string or function. If a string, then it is a
\code{sprintf} in which the 1st pattern will be replaced with the variable
index number. If a function, then it is a function that takes the index
nubmer as its first argument.}

\item{...}{Additional arguments to pass to \code{.f}}
}
\description{
Rename tables with sequentially numbered variables
}
