% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_sub}
\alias{rename_sub}
\title{Rename tables with a regular expression}
\usage{
rename_sub(.data, pattern, replacement, all = TRUE)
}
\arguments{
\item{.data}{A data frame}

\item{pattern, replacement}{Pattern and replacement regular expressions.
See \code{\link[stringr]{str_replace}}.}

\item{all}{If \code{TRUE}, the replace all patterns, else replace only the
the first pattern.}
}
\description{
Use a regular expression to rename tables.
The function \code{rename_replace} applies \code{\link[stringr]{str_replace}} to the names of the table,
while \code{rename_replace_all} applies \code{\link[stringr]{str_replace_all}}.
}
