% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_rows.R
\name{rep_rows}
\alias{rep_rows}
\title{Replicate rows of a data frame}
\usage{
rep_rows(.data, times = 1, each = NA)
}
\arguments{
\item{.data}{A data frame}

\item{times}{See \code{\link{rep}}}

\item{each}{See \code{\link{rep}}}
}
\value{
A data frame with replicated rows
}
\description{
Replicate rows of a data frame
}
\examples{
require("tibble")
x <- tibble(a = 1:2, b = 3:4)
rep_rows(x, 2)
rep_rows(x, each = 2)
}

