% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_names.R
\name{replace_names}
\alias{replace_names}
\title{Replace names in a vector}
\usage{
replace_names(x, nm)
}
\arguments{
\item{x}{Vector to name}

\item{nm}{Named character vector, where the names are the
the new names and the values are the old names.}
}
\value{
The vector \code{x} with new names.
}
\description{
Unlike \code{\link[purrr]{set_names}} changes a subset
of names.
}
\examples{
replace_names(c(a = 1, b = 2), c("beta" = "b"))
}

