% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_x_log2}
\alias{scale_x_log2}
\alias{scale_y_log2}
\title{Log2 scale transformation}
\usage{
scale_x_log2(...)

scale_y_log2(...)
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2]{scale_x_continuous}}}
}
\description{
Constructs log 2 continuous scales for \code{ggplot2}, similar
to \code{\link[ggplot2]{scale_x_log10}}.
}
