% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcor2cov.R
\name{sdcor2cov}
\alias{sdcor2cov}
\title{Standard deviations and Correlation to Covariance Matrix}
\usage{
sdcor2cov(s, r = diag(length(s)))
}
\arguments{
\item{s}{A vector of standard deviations}

\item{r}{A correlation matrix with \code{dim(length(s), length(x)}.}
}
\value{
A \code{dim(length(s), length(s))} covariance matrix.
}
\description{
Standard deviations and Correlation to Covariance Matrix
}
