% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junk.R
\name{set_names_seq}
\alias{set_names_seq}
\title{Set names in a vector with sequential numbering}
\usage{
set_names_seq(x, .f = "Var\%d", ...)
}
\arguments{
\item{x}{An object to be named}

\item{.f}{A character vector or function. If \code{.f} is a function,
then its first argument will be the index numbers of the names.
If \code{.f} is a character, then it should be a format string
for \code{sprintf} pattern.}

\item{...}{Arguments passed to \code{.f}.}
}
\value{
The object \code{x} with its names (colnames, rownames) set.
}
\description{
Set the names in vector with a pattern based on sequential numbering,
e.g. \code{Var1}, \code{Var2}, ...
}
\seealso{
\code{\link[stats]{setNames}}, \code{\link[purrr]{set_names}},
  \code{\link[magrittr]{set_colnames}}, \code{\link[magrittr]{set_colnames}}.
}

