% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_coef}
\alias{standardize_coef}
\alias{unstandardize_coef}
\title{Standardize and Unstandardize Coefficients}
\usage{
standardize_coef(beta, x_mean = 0, x_sd = 1, y_mean = NULL, y_sd = NULL)

unstandardize_coef(beta, x_mean = 0, x_sd = 1, y_mean = NULL,
  y_sd = NULL)
}
\arguments{
\item{beta}{numeric vector with length \code{K} of the the coefficients.}

\item{x_mean}{numeric vector of length \code{K - 1}, with the means of the associated variables.}

\item{x_sd}{numeric vector of length \code{K - 1}, with the standard deviations of the associated variables.}

\item{y_mean}{number with the mean of the response variable.}

\item{y_sd}{number with the standard deviation of the response variable.}
}
\value{
A numeric vector of length \code{K}.
}
\description{
Given a vector of \eqn{K} coeficients (\eqn{\beta}{beta}), and the means (\eqn{\bar{x}}{mean(x)}) and standard deviations (\eqn{s_x}(sd(x))) of X, \code{standardize} standardizes them.
The intercept is,
\deqn{
\beta_0^* = \frac{\beta_0 - \sum_{k = 1}^K \beta_k \frac{\bar{x_k}}{s_x} - \bar{y}}{s_y}
}{
beta_0^* = (beta_0 - sum(beta * x / sd(x) - mean(y))) / sd(y)
}
and the other coefficients are,
\deqn{
beta_k^* = \frac{\beta_k}{s_x s_y},
}{beta^* = beta / (sd(x) * sd(y)),}
for \eqn{k = 1, \dots, K}{k = 1, ..., K}.
}
\details{
The function \code{unstandardize} converts standardized coefficients back
to their unstandardized forms.
For the intercept, the conversion is
\deqn{
\beta_0 = s_y \left(\beta_0 + \sum_{k} \beta_k^*  \times  mean(x_k) \times sd(x_k) \right) + \bar{y} .
}{
beta^* = sd_y * (beta + sum(beta * mean(x) * sd(x))) + mean(y)
}
For the other coefficients, the conversion is,
\deqn{
\beta_k = \beta_k^* s_x s_y
}{beta^* = beta * sd(x) * sd(y)}
}
