% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummies.R
\name{to_dummies}
\alias{to_dummies}
\alias{to_dummies_}
\title{Convert column to dummy variables}
\usage{
to_dummies(.data, col, remove = TRUE, drop_level = FALSE, vars = NULL)

to_dummies_(.data, col, remove = TRUE, drop_level = FALSE, vars = NULL)
}
\arguments{
\item{.data}{A data frame}

\item{col}{The bare column name}

\item{remove}{If \code{TRUE}, then remove the input column from the output
data frame.}

\item{drop_level}{If \code{TRUE}, then drop a level when creating the variables.
This is the default for models.}

\item{vars}{A \code{character} vector with the names of the new variables,
or a function with two arguments, the column name and the factor levels, which
returns a character vector of new variable names.}
}
\value{
A data frame.
}
\description{
\code{to_dummies} converts a column into multiple columns of dummy variables.
\code{to_dummies_} is the standard evaluation version of this function.
}
\examples{
example_df <- data.frame(x = c(rep("a", 2), rep("b", 2)))
to_dummies(example_df, x)
to_dummies(example_df, x, remove = FALSE)
to_dummies(example_df, x, drop_level = TRUE)
}

