% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncare.R
\name{truncare}
\alias{truncare}
\alias{censor}
\title{Truncate or censor vectors}
\usage{
truncare(x, left = -Inf, right = Inf, newleft = NA, newright = NA,
  leq = FALSE, geq = FALSE)

censor(x, left = -Inf, right = Inf)
}
\arguments{
\item{x}{A vector}

\item{left, right}{Boundary values for truncation}

\item{newleft, newright}{Value to replace truncated elements}

\item{leq, geq}{If \code{TRUE} use \code{<=} (\code{>=}) for tests at the boundary,
otherwise use \code{<} (\code{>}).}
}
\value{
Vector of the same size as \code{x}, with truncated elements replaced.
}
\description{
\code{truncare} will truncate a vector with replacing elements greater or less than
given bounds with replacement values (missing by default).
\code{censor} is a special case that uses the upper and bound values as the
replacements.
}
\details{
The more intuitive name \code{truncate} was not chosen, because a base function \code{\link[base]{truncate}}
already has that name. \emph{Truncare} is the Latin origin of truncate.
}
\examples{
truncare(1:10, left = 2, right = 5)
censor(-5:5, 0, 4)
}
