% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate.R
\name{truncate}
\alias{censor}
\alias{truncate}
\title{Truncate or censor vectors}
\usage{
truncate(x, left = -Inf, right = Inf, newleft = NA, newright = NA,
  leq = FALSE, geq = FALSE)

censor(x, left = -Inf, right = Inf)
}
\arguments{
\item{x}{A vector}

\item{left, right}{Boundary values for truncation}

\item{newleft, newright}{Value to replace truncated elements}

\item{leq, geq}{If \code{TRUE} use \code{<=} (\code{>=}) for tests at the boundary,
otherwise use \code{<} (\code{>}).}
}
\value{
Vector of the same size as \code{x}, with truncated elements replaced.
}
\description{
\code{truncate} will truncate a vector with replacing elements greater or less than
given bounds with replacement values (missing by default).
\code{censor} is a special case that uses the upper and bound values as the
replacements.
}

