% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfill.R
\name{unfill}
\alias{unfill}
\title{Unfill a vector}
\usage{
unfill(x, tokeep = c("first", "last", "middle"), value = NA)
}
\arguments{
\item{x}{A vector}

\item{tokeep}{If \code{"first"}, then keep the first value. If \code{"last"},
then keep the last value in the run. If \code{"middle"}, keep the middle
value.}

\item{value}{The value to use to unfill the vector.}
}
\value{
The vector \code{x} with the elements other than those
  specified in \code{tokeep}, replaced with \code{value}.
}
\description{
Unfill a vector
}

