% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{fct_remove}
\alias{fct_remove}
\title{Remove levels from a factor}
\usage{
fct_remove(f, lvls)
}
\arguments{
\item{f}{A factor}

\item{lvls}{Character vector of levels to remove}
}
\value{
A factor
}
\description{
Remove levels from a factor, meaning that observations with those levels
are set to `NA`.
}
\examples{
f <- factor(c("Low", "Medium", "High",
              "Refused to respond", "No response", "Not asked"))
fct_remove(f, c("Refused to respond", "No response", "Not asked"))
}
\seealso{
\code{\link[forcats]{fct_recode}} which can
  remove factor levels and \code{\link[forcats]{fct_explicit_na}} which
  is the inverse, converting `NA` to a factor level.
}
