% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_sub}
\alias{rename_sub}
\title{Rename tables with a regular expression}
\usage{
rename_sub(.data, pattern, replacement, all = TRUE)
}
\arguments{
\item{.data}{A data frame}

\item{pattern, replacement}{Pattern and replacement regular expressions.
See \code{\link[stringr]{str_replace}}.}

\item{all}{If \code{TRUE}, then use \code{str_replace_all} to replace
names, else use \code{str_replace}.}
}
\description{
Rename tables with a regular expression
}

