% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_rows.R
\name{rep_rows}
\alias{rep_rows}
\title{Replicate rows of a data frame}
\usage{
rep_rows(.data, times = NA, each = NA)
}
\arguments{
\item{.data}{A data frame}

\item{times}{See \code{\link{rep}}}

\item{each}{See \code{\link{rep}}}
}
\value{
A data frame with replicated rows
}
\description{
Replicate rows of a data frame
}

