% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_names.R
\name{replace_names}
\alias{replace_names}
\title{Set names in a vector by regular expression}
\usage{
replace_names(x, pattern, replacement, all = TRUE)
}
\arguments{
\item{x}{Vector to name}

\item{pattern, replacement}{Pattern and replacement strings. See \code{\link[stringr]{str_replace}}}

\item{all}{Use \code{\link[stringr]{str_replace_all}},
else use \code{\link[stringr]{str_replace}}.}
}
\value{
Return table with renamed variables
}
\description{
Set names in a vector by regular expression
}
\examples{
replace_names(c(a = 1, b = 2), c("beta" = "b"))
fruits <- c("one apple" = 1, "two pears" = 2, "three bananas" = 3)
replace_names(fruits, "[aeiou]", "-")
replace_names(fruits, "[aeiou]", "-", all = TRUE)
}

