% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_col.R
\name{select_col}
\alias{select_col}
\alias{select_col_}
\title{Select single column from a table}
\usage{
select_col(.data, col, nm = NULL)

select_col_(.data, col, nm = NULL)
}
\arguments{
\item{.data}{A data frame}

\item{col}{In \code{select_col} an unquoted
column name. In \code{select_col_}, the
column name as a string.}

\item{nm}{If non-\code{NULL}, names to apply to the vector.
In \code{select_col} an unquoted column name.
In \code{select_col_}, the column name as a string.}
}
\description{
Select a single column from a table and
return as a vector. This is pretty-much
just a wrapper for `[[`, but is more
descriptive and pipe friendly.
}
\details{
If the \code{nm} argument is used, then this
can be seen as the inverse of \code{\link[tibble]{enframe}}.
}
\examples{
select_col(iris, Species)
select_col_(iris, "Species")
}
