% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_names.R
\name{set_names_map}
\alias{set_names_map}
\title{Set names in a vector with a function}
\usage{
set_names_map(x, .f, ...)
}
\arguments{
\item{x}{Vector to name}

\item{.f}{Function to use to name the variables.
This is converted to a function using \code{\link[purrr]{as_function}}.
The first argument will be \code{names(x)}, pass additional functions with
\code{...}. It should return a character vector of the same length as the
names of \code{x}.}

\item{...}{Arguments passed to the function \code{.f}.}
}
\value{
The vector \code{x} with new names
}
\description{
Set names in a vector by applying a function to the original names.
}

