% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_names.R
\name{set_names_sub}
\alias{set_names_sub}
\title{Set names in a vector by regular expression}
\usage{
set_names_sub(x, pattern, replacement, all = TRUE)
}
\arguments{
\item{x}{Vector to name}

\item{pattern, replacement}{Pattern and replacement strings. See \code{\link[stringr]{str_replace}}}

\item{all}{Use \code{\link[stringr]{str_replace_all}},
else use \code{\link{str_replace}}.}
}
\value{
Return table with renamed variables
}
\description{
Set names in a vector by regular expression
}

