% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummies.R
\name{to_dummies}
\alias{to_dummies}
\alias{to_dummies_}
\title{Convert column to dummy variables}
\usage{
to_dummies(.data, col, sep = "_", exclude = NULL)

to_dummies_(.data, col, sep = "_", exclude = NULL)
}
\arguments{
\item{.data}{A tbl.}

\item{col}{In \code{to_dummies}, the bare (unquoted) column name.
In \code{to_dummies_}, a character vector with column name.}

\item{sep}{The new variable names are constructed from \code{col} and the
level value separated by \code{sep}.}

\item{exclude}{Ignore these levels of \code{col} when creating variables.}
}
\value{
A tbl
}
\description{
\code{to_dummies} converts a column into multiple columns of dummy variables.
\code{to_dummies_} is the standard evaluation version of this function.
}

